levels/08-markdown-reconstruction/requirements.txt
pymupdf>=1.23.0
pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
